requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
httpx==0.27.0
Pillow==9.5.0
simple-lama-inpainting==0.1.2
numpy==1.26.4
python-multipart==0.0.9
pytest==8.3.2
